remove iinv, add adapter
