+++ src/main/java/ruiseki/okbackpack/compat/structurelib/StructureLibCompat.java
@@ -2,7 +2,6 @@ package ruiseki.okbackpack.compat.structurelib;
 
 import java.util.function.Function;
 
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.IInventory;
 import net.minecraft.item.ItemStack;
 
@@ -13,6 +12,7 @@ import ruiseki.okbackpack.OKBackpack;
 import ruiseki.okbackpack.common.block.BackpackWrapper;
 import ruiseki.okbackpack.common.block.BlockBackpack;
 import ruiseki.okbackpack.common.helpers.BackpackEntityHelpers;
+import ruiseki.okbackpack.common.inventory.BackpackWrapperInventoryAdapter;
 import ruiseki.okbackpack.compat.Mods;
 import ruiseki.okcore.init.IInitListener;
 
@@ -44,85 +44,7 @@ public class StructureLibCompat implements IInitListener {
                 return null;
             }
             BackpackWrapper wrapper = new BackpackWrapper(stack, (BlockBackpack.ItemBackpack) stack.getItem());
-            return new StructureLibBackpackInventory(wrapper);
-        }
-    }
-
-    public static final class StructureLibBackpackInventory implements IInventory {
-
-        public final BackpackWrapper wrapper;
-
-        public StructureLibBackpackInventory(BackpackWrapper wrapper) {
-            this.wrapper = wrapper;
-        }
-
-        @Override
-        public int getSizeInventory() {
-            return wrapper.getSlots();
-        }
-
-        @Override
-        public ItemStack getStackInSlot(int slot) {
-            return wrapper.getStackInSlot(slot);
-        }
-
-        @Override
-        public ItemStack decrStackSize(int slot, int amount) {
-            return wrapper.extractItem(slot, amount, false);
-        }
-
-        @Override
-        public ItemStack getStackInSlotOnClosing(int slot) {
-            ItemStack stack = wrapper.getStackInSlot(slot);
-            if (stack != null) {
-                wrapper.setStackInSlot(slot, null);
-            }
-            return stack;
-        }
-
-        @Override
-        public void setInventorySlotContents(int slot, ItemStack stack) {
-            wrapper.setStackInSlot(slot, stack);
-        }
-
-        @Override
-        public String getInventoryName() {
-            return wrapper.getInventoryName();
-        }
-
-        @Override
-        public boolean hasCustomInventoryName() {
-            return wrapper.hasCustomInventoryName();
-        }
-
-        @Override
-        public int getInventoryStackLimit() {
-            return Integer.MAX_VALUE;
-        }
-
-        @Override
-        public void markDirty() {
-            wrapper.markDirty();
-            wrapper.writeToItem();
-        }
-
-        @Override
-        public boolean isUseableByPlayer(EntityPlayer player) {
-            return true;
-        }
-
-        @Override
-        public void openInventory() {}
-
-        @Override
-        public void closeInventory() {
-            wrapper.writeToItem();
-        }
-
-        @Override
-        public boolean isItemValidForSlot(int slot, ItemStack stack) {
-            return wrapper.getStackHandler()
-                .isItemValid(slot, stack);
+            return new BackpackWrapperInventoryAdapter(wrapper);
         }
     }
 }
